+++ framework/packages/abstract-sdk/src/base/features/identification.rs
@@ -39,21 +39,6 @@ pub trait AccountIdentification: Sized {
     }
 }
 
-/// Trait for modules that allowed to execute on proxy
-pub trait AccountExecutor: AccountIdentification {
-    /// Execute proxy method on proxy contract
-    fn execute_on_proxy(
-        &self,
-        deps: Deps,
-        msg: &proxy::ExecuteMsg,
-    ) -> AbstractSdkResult<CosmosMsg> {
-        let proxy_address = self.proxy_address(deps)?;
-        wasm_execute(proxy_address, msg, vec![])
-            .map(Into::into)
-            .map_err(Into::into)
-    }
-}
-
 #[cfg(test)]
 mod test {
     use abstract_testing::prelude::*;
