+++ framework/packages/abstract-sdk/src/base/features/mod.rs
@@ -2,12 +2,13 @@ mod abstract_name_service;
 mod abstract_registry_access;
 mod dependencies;
 mod identification;
+mod executor;
 mod module_identification;
 
 pub use abstract_name_service::{AbstractNameService, AbstractNameServiceClient};
 pub use abstract_registry_access::AbstractRegistryAccess;
 pub use dependencies::Dependencies;
-pub use identification::AccountExecutor;
+pub use executor::AccountExecutor;
 pub use identification::AccountIdentification;
 pub use module_identification::ModuleIdentification;
